requirements.txt
torch
torchvision
matplotlib
pillow
scikit-learn
numpy